+++ src/Controller/UserController.php
@@ -13,6 +13,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
  * @Route("/users")
@@ -24,6 +25,10 @@ class UserController extends AbstractController
      */
     public function index(UserRepository $accountsRepository, Request $request, PaginatorInterface $paginator): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('You do not have access to this page.');
+        }
+
         $data = $accountsRepository->findAll();
         // Paginate the data
         $pagination = $paginator->paginate(
@@ -41,6 +46,10 @@ class UserController extends AbstractController
      */
     public function new(Request $request, UserRepository $accountsRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('You do not have access to this page.');
+        }
+
         $account = new User();
         $account->setAddedOn(new \DateTime());
         $form = $this->createForm(UserType::class, $account);
@@ -80,6 +89,10 @@ class UserController extends AbstractController
      */
     public function show(User $account): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('You do not have access to this page.');
+        }
+
         return $this->render('users/show.html.twig', [
                 'account' => $account,
         ]);
@@ -90,6 +103,10 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $account, UserRepository $accountsRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('You do not have access to this page.');
+        }
+
         $form = $this->createForm(UserType::class, $account);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -127,6 +144,10 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $account, UserRepository $accountsRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('You do not have access to this page.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$account->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($account);
